// src/store/user/user.types.ts
export enum USER_TYPES {
  ADD_CURRENT_REDUCER = "ADD_CURRENT_USER",
  CHECK_USER_SESSION = "CHCECK_USER_SESSION",
  GOOGLE_SIGNIN_START = "GOOGLE_SIGNIN_START",
  EMAIL_SIGNIN_START = "EMAIL_SIGNIN_START",
  SIGNIN_SUCCESS = "SIGNIN_SUCCESS",
  SIGNIN_FAILED = "SIGNIN_FAILED",
  SIGNUP_START = "SIGNUP_START",
  SIGNUP_SUCCESS = "SIGNUP_SUCCESS",
  SIGNUP_FAILED = "SIGNUP_FAILED",
};